+++ app/guid-node/registrations/controller.ts
@@ -42,7 +42,7 @@ export default class GuidNodeRegistrations extends Controller {
         let schemas = yield this.store.findAll('registration-schema');
         schemas = schemas.toArray();
         schemas.sort((a: RegistrationSchema, b: RegistrationSchema) => {
-            return a.name.length > b.name.length;
+            return a.name.length - b.name.length;
         });
         this.set('defaultSchema', schemas.firstObject);
         this.set('selectedSchema', this.defaultSchema);
@@ -56,9 +56,9 @@ export default class GuidNodeRegistrations extends Controller {
         return this.tab ? this.tab : 'registrations';
     }
 
-    @computed('node.id', 'node.root.id', 'node.root.currentUserIsAdmin')
+    @computed('node.id', 'node.root.id', 'node.root.userHasAdminPermission')
     get isComponentRootAdmin() {
-        return this.node && this.node.id !== this.node.root.get('id') && this.node.root.get('currentUserIsAdmin');
+        return this.node && this.node.id !== this.node.root.get('id') && this.node.root.get('userHasAdminPermission');
     }
 
     @action