+++ app/locales/en/translations.ts
@@ -58,6 +58,12 @@ export default {
         required: 'Required',
         options: 'Options',
         optional: 'Optional',
+        services: {
+            collections: 'Collections',
+            institutions: 'Institutions',
+            preprints: 'Preprints',
+            registries: 'Registries',
+        },
     },
     maintenance: {
         line1: 'The site will undergo maintenance between <strong>{{start}} and {{end}}</strong> ({{utc}} UTC).',
@@ -252,6 +258,8 @@ export default {
         toggle_primary: 'Toggle primary navigation',
         toggle_secondary: 'Toggle secondary navigation',
         other_views: 'Other OSF views',
+        login: 'Login',
+        join: 'Join',
     },
     auth_dropdown: {
         log_out: 'Log Out',
@@ -1000,6 +1008,11 @@ export default {
                 more: 'See more',
             },
         },
+
+        overview: {
+            overview: 'Overview',
+            metadata: 'Metadata',
+        },
     },
     analytics: {
         pageTitle: '{{nodeTitle}} Analytics',
@@ -1197,5 +1210,32 @@ export default {
                 editScopes: 'Edit scopes',
             },
         },
+    },
+    verifyEmail: {
+        merge: {
+            header: 'Merge account?',
+            body: 'Would you like to merge <strong>{{email}}</strong> into your account? This action is irreversible.',
+            verifyButton: 'Merge account',
+            denyButton: 'Do not merge account',
+            verifySuccess: '<strong>{{email}}</strong> has been merged into your account.',
+            denySuccess: '<strong>{{email}}</strong> has not been merged into your account.',
+            verifyError: 'There was a problem merging <strong>{{email}}</strong> into your account.',
+            denyError: 'There was a problem canceling the request to merge <strong>{{email}}</strong> into your account.',
+        },
+        add: {
+            header: 'Add alternate email?',
+            body: 'Would you like to add <strong>{{email}}</strong> to your account?',
+            verifyButton: 'Add email',
+            denyButton: 'Do not add email',
+            verifySuccess: '<strong>{{email}}</strong> has been added to your account.',
+            denySuccess: '<strong>{{email}}</strong> has not been added to your account.',
+            verifyError: 'There was a problem adding <strong>{{email}}</strong> to your account.',
+            denyError: 'There was a problem canceling the request to add <strong>{{email}}</strong> to your account.',
+        },
+    },
+    ipsum: {
+        title: 'tempor nec feugiat nisl pretium',
+        sentence: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+        paragraph: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Id velit ut tortor pretium. Nisi porta lorem mollis aliquam ut porttitor leo a. Cras fermentum odio eu feugiat. Eget mi proin sed libero enim. Quam adipiscing vitae proin sagittis. Volutpat consequat mauris nunc congue nisi vitae suscipit tellus. At varius vel pharetra vel turpis nunc eget. Purus ut faucibus pulvinar elementum integer enim neque volutpat. Turpis nunc eget lorem dolor. Mattis pellentesque id nibh tortor id aliquet lectus proin nibh. Arcu felis bibendum ut tristique et egestas quis. Nisl tincidunt eget nullam non nisi est sit amet. Fringilla urna porttitor rhoncus dolor purus non enim.',
     },
 };