+++ lib/osf-components/addon/components/validated-model-form/component.ts
@@ -2,6 +2,8 @@ import { layout } from '@ember-decorators/component';
 import { or } from '@ember-decorators/object/computed';
 import { service } from '@ember-decorators/service';
 import Component from '@ember/component';
+import { assert } from '@ember/debug';
+import { set } from '@ember/object';
 import { task } from 'ember-concurrency';
 import { Validations } from 'ember-cp-validations';
 import DS, { ModelRegistry } from 'ember-data';
@@ -23,6 +25,7 @@ export default class ValidatedModelForm<M extends ValidatedModelName> extends Co
     @requiredAction onSave!: (model: ModelRegistry[M]) => void;
 
     // Optional arguments
+    onError?: (e: object, model: ModelRegistry[M]) => void;
     model?: ModelRegistry[M];
     modelName?: M; // If provided, new model instance created in constructor
     disabled: boolean = defaultTo(this.disabled, false);
@@ -35,6 +38,7 @@ export default class ValidatedModelForm<M extends ValidatedModelName> extends Co
 
     shouldShowMessages: boolean = false;
     saved: boolean = false;
+    modelProperties: object = defaultTo(this.modelProperties, {});
 
     @or('disabled', 'saveModelTask.isRunning')
     inputsDisabled!: boolean;
@@ -44,7 +48,7 @@ export default class ValidatedModelForm<M extends ValidatedModelName> extends Co
             return;
         }
         if (this.analyticsScope) {
-            this.analytics.click('button', `${this.analyticsScope} - Save developer app`);
+            this.analytics.click('button', `${this.analyticsScope} - Save`);
         }
 
         const { validations } = yield this.model.validate();
@@ -55,8 +59,16 @@ export default class ValidatedModelForm<M extends ValidatedModelName> extends Co
                 yield this.model.save();
                 this.set('saved', true);
                 this.onSave(this.model);
+                if (this.modelName) {
+                    set(this, 'model', this.store.createRecord(this.modelName, this.modelProperties));
+                }
+                this.set('shouldShowMessages', false);
             } catch (e) {
-                this.toast.error(e);
+                if (this.onError) {
+                    this.onError(e, this.model);
+                } else {
+                    this.toast.error(e);
+                }
                 throw e;
             }
         }
@@ -65,8 +77,10 @@ export default class ValidatedModelForm<M extends ValidatedModelName> extends Co
     constructor(...args: any[]) {
         super(...args);
 
+        assert('Can only pass either a model or a modelName', !(Boolean(this.model) && Boolean(this.modelName)));
+
         if (!this.model && this.modelName) {
-            this.model = this.store.createRecord(this.modelName, {});
+            this.model = this.store.createRecord(this.modelName, this.modelProperties);
         }
     }