+++ lib/registries/addon/overview/controller.ts
@@ -1,0 +1,42 @@
+import { action, computed } from '@ember-decorators/object';
+import { not } from '@ember-decorators/object/computed';
+import { service } from '@ember-decorators/service';
+import Controller from '@ember/controller';
+import Media from 'ember-responsive';
+
+export default class Overview extends Controller {
+    @service media!: Media;
+    @not('media.isDesktop') showMobileNav!: boolean;
+
+    sidenavGutterClosed = true;
+    metadataGutterClosed = true;
+
+    @computed('media.{isMobile,isTablet,isDesktop}')
+    get metadataGutterMode() {
+        if (this.media.isMobile) {
+            return 'page';
+        }
+        if (this.media.isTablet) {
+            return 'drawer';
+        }
+        return 'column';
+    }
+
+    @computed('media.{isMobile,isTablet,isDesktop}')
+    get sidenavGutterMode() {
+        if (this.media.isDesktop) {
+            return 'column';
+        }
+        return 'drawer';
+    }
+
+    @action
+    toggleSidenav() {
+        this.toggleProperty('sidenavGutterClosed');
+    }
+
+    @action
+    toggleMetadata() {
+        this.toggleProperty('metadataGutterClosed');
+    }
+}